python/pyproject.toml
[tool.poetry]
name = "e2b-desktop"
version = "0.0.1"
description = "E2B Desktop - Desktop in cloud powered by E2B"
authors = ["e2b <hello@e2b.dev>"]
license = "Apache-2.0"
homepage = "https://e2b.dev/"
repository = "https://github.com/e2b-dev/desktop"
packages = [{ include = "e2b_desktop" }]

[tool.poetry.dependencies]
python = "^3.8"

pydantic = ">1, <3"
e2b = "^1.0.1"
pytesseract = "^0.3.13"
pillow = "10.4.0"
easyocr = "^1.7.2"
numpy = "1.24"
opencv-python = "4.8.1.78"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pytest = "^7.4.0"
python-dotenv = "^1.0.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/e2b-dev/desktop/issues"
